% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdal0-package.R
\docType{package}
\name{rmdal0-package}
\alias{rmdal0}
\alias{rmdal0-package}
\title{rmdal0: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
