% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdal0-package.R
\docType{package}
\name{rmdal0-package}
\alias{rmdal0}
\alias{rmdal0-package}
\title{rmdal0: Bindings to the Mesh Data Abstraction Library (MDAL)}
\description{
Mesh stuff with MDAL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/rmdal0}
  \item Report bugs at \url{https://github.com/mdsumner/rmdal0/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
